finetune/pyproject.toml
[project]
name = "finetune"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "numpy>=2.2.6",
    "pandas>=2.3.2",
    "pillow>=12.0.0",
    "protobuf>=6.33.0",
    "scikit-learn>=1.7.2",
    "sentencepiece>=0.2.1",
    "torch>=2.9.1",
    "torchvision>=0.24.1",
    "tqdm>=4.67.1",
    "transformers>=4.57.1",
]
